// src/enums/currently-playing-IPCChannels.enum.ts
export enum CurrentlyPlayingIPCChannels {
  SetCurrentVideo = "set-current-video-backend",
  SET_CURRENTLY_PLAYING_CURRENTTIME = "set-currently-playing-currenttime-backend",
  SetCurrentPlaylist = "set-current-playlist-backend",
  GetCurrentPlaylist = "get-current-playlist-backend",
  GetPlaylistVideos = "get-playlist-videos-backend",
  GetNextPlaylistVideo = "get-next-playlist-video-backend",
  GetPreviousPlaylistVideo = "get-previous-playlist-video-backend",
  GetCurrentVideo = "get-current-video-backend",
}